requirements.txt
django>=4.1.1
djangorestframework
pymysql
python-dotenv
uvicorn
gunicorn
uvloop; platform_system != 'Windows'
httptools
